Move testRollup constants to public/utils
