+++ public/pages/RollupDetails/containers/RollupDetails/RollupDetails.test.tsx
@@ -19,14 +19,14 @@ import { toastNotifications } from "ui/notify";
 import { render, wait } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 import { MemoryRouter as Router } from "react-router";
+import { Redirect, Route, RouteComponentProps, Switch } from "react-router-dom";
 import { browserServicesMock } from "../../../../../test/mocks";
 import { BrowserServices } from "../../../../models/interfaces";
 import { ModalProvider, ModalRoot } from "../../../../components/Modal";
-import { Redirect, Route, RouteComponentProps, Switch } from "react-router-dom";
 import { BREADCRUMBS, ROUTES } from "../../../../utils/constants";
 import RollupDetails from "./RollupDetails";
-import { testRollup, testRollup2 } from "../../../CreateRollup/utils/constants";
 import { ServicesConsumer, ServicesContext } from "../../../../services";
+import { testRollup } from "../../../../utils/constants.test";
 
 function renderRollupDetailsWithRouter(initialEntries = ["/"]) {
   return {
